working, split up into files
