+++ constants.py
@@ -1,0 +1,4 @@
+import os
+
+# Define the current directory as a constant
+CUR_DIR = os.path.dirname(os.path.abspath(__file__))
