+++ yolo_funcs.py
@@ -1,0 +1,74 @@
+import cv2
+import numpy as np
+import os
+import urllib.request
+
+from constants import CUR_DIR
+
+
+# Download YOLO model files if they do not exist
+def download_yolo_files():
+    yolo_dir = "/Users/Henry/Desktop/github/cv-soccer/yolo"
+    if not os.path.exists(yolo_dir):
+        os.makedirs(yolo_dir)
+    
+    files = {
+        "yolov3.weights": "https://pjreddie.com/media/files/yolov3.weights",
+        "yolov3.cfg": "https://raw.githubusercontent.com/pjreddie/darknet/master/cfg/yolov3.cfg",
+        "coco.names": "https://raw.githubusercontent.com/pjreddie/darknet/master/data/coco.names"
+    }
+    
+    for file_name, url in files.items():
+        file_path = os.path.join(yolo_dir, file_name)
+        if not os.path.exists(file_path):
+            print(f"Downloading {file_name}...")
+            urllib.request.urlretrieve(url, file_path)
+            print(f"Downloaded {file_name}")
+
+# Load YOLO model
+def load_yolo_model():
+    net = cv2.dnn.readNet(os.path.join(CUR_DIR, "yolo/yolov3.weights"), os.path.join(CUR_DIR, "yolo/yolov3.cfg"))
+    with open(os.path.join(CUR_DIR, "yolo/coco.names"), "r") as f:
+        classes = [line.strip() for line in f.readlines()]
+    layer_names = net.getLayerNames()
+    output_layers = [layer_names[i - 1] for i in net.getUnconnectedOutLayers()]
+    return net, classes, output_layers
+
+
+# Run inference on video frames
+def detect_humans(frame, net, classes, output_layers):
+    height, width, channels = frame.shape
+    blob = cv2.dnn.blobFromImage(frame, 0.00392, (416, 416), (0, 0, 0), True, crop=False)
+    net.setInput(blob)
+    outs = net.forward(output_layers)
+
+    class_ids = []
+    confidences = []
+    boxes = []
+
+    for out in outs:
+        for detection in out:
+            scores = detection[5:]
+            class_id = np.argmax(scores)
+            confidence = scores[class_id]
+            if confidence > 0.5 and class_id == 0:  # Class ID 0 is for 'person' in COCO dataset
+                center_x = int(detection[0] * width)
+                center_y = int(detection[1] * height)
+                w = int(detection[2] * width)
+                h = int(detection[3] * height)
+                x = int(center_x - w / 2)
+                y = int(center_y - h / 2)
+                boxes.append([x, y, w, h])
+                confidences.append(float(confidence))
+                class_ids.append(class_id)
+
+    indexes = cv2.dnn.NMSBoxes(boxes, confidences, 0.5, 0.4)
+    for i in range(len(boxes)):
+        if i in indexes:
+            x, y, w, h = boxes[i]
+            label = str(classes[class_ids[i]])
+            cv2.rectangle(frame, (x, y), (x + w, y + h), (0, 255, 0), 2)
+            cv2.putText(frame, label, (x, y - 10), cv2.FONT_HERSHEY_SIMPLEX, 0.5, (0, 255, 0), 2)
+
+    return frame
+
